refactor: change structure of the project, remove unnecessary dependencies
