--- remote-control/.env
@@ -1,1 +1,0 @@
-PORT=8080
