+++ remote-control/package.json
@@ -32,6 +32,7 @@
     "@typescript-eslint/parser": "^5.45.0",
     "babel-loader": "^9.1.2",
     "clean-webpack-plugin": "^4.0.0",
+    "cross-env": "^7.0.3",
     "eslint": "^8.28.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
@@ -48,7 +49,6 @@
   },
   "dependencies": {
     "@nut-tree/nut-js": "^2.3.0",
-    "dotenv": "^16.0.3",
     "jimp": "^0.16.2",
     "ws": "^8.11.0"
   }
