+++ remote-control/src/index.ts
@@ -1,0 +1,68 @@
+import { createWebSocketStream } from 'ws';
+import { EOL } from 'os';
+import messageHandler from './utils/messageHandler';
+import httpServer from './http_server/index';
+import wss from './ws_server/index';
+
+const WS_PORT = 8080;
+const HTTP_PORT = 8181;
+
+httpServer.listen(HTTP_PORT);
+
+process.stdout.write(
+  `\x1b[36mhttp server started on http://localhost:${HTTP_PORT}${EOL}`
+);
+
+wss.on('listening', () =>
+  process.stdout.write(
+    `\x1b[36mWS server started on http://localhost:${WS_PORT}${EOL}`
+  )
+);
+
+wss.on('headers', (data) => {
+  process.stdout.write(`\x1b[91mweb socket started ${EOL}`);
+  data.forEach((element: string) => {
+    process.stdout.write(`\x1b[33m${element} ${EOL}`);
+  });
+});
+
+wss.on('connection', (ws) => {
+  try {
+    const duplex = createWebSocketStream(ws, {
+      encoding: 'utf8',
+      decodeStrings: false,
+    });
+
+    duplex.write(`connection_is_on_localhost:${WS_PORT}`);
+
+    duplex.on('readable', async () => {
+      let data = '';
+      let chunk = '';
+      while (chunk !== null) {
+        data += chunk;
+        chunk = duplex.read();
+      }
+      if (data) {
+        process.stdout.write(`\x1b[34mWS received the command ${data} ${EOL}`);
+      }
+
+      const message = await messageHandler(data, duplex);
+
+      if (message) {
+        process.stdout.write(`\x1b[96mWS send the command ${message} ${EOL}`);
+      }
+    });
+
+    ws.on('close', () => {
+      process.stdout.write(`\x1b[91mwebsocket closed ${EOL} ${EOL}`);
+      duplex.destroy();
+    });
+  } catch {
+    process.stdout.write('Server error');
+  }
+});
+
+process.on('SIGINT', () => {
+  process.stdout.write(`\x1b[37m`);
+  process.exit();
+});
